Use database page size in read replication
